feat: add styling to components
