+++ src/App.css
@@ -1,8 +1,9 @@
 #root {
   max-width: 1280px;
   margin: 0 auto;
-  padding: 2rem;
+  padding: 1rem;
   text-align: center;
+  min-width: 0;
 }
 
 .logo {
@@ -39,4 +40,21 @@
 
 .read-the-docs {
   color: #888;
+}
+
+h1,
+h2 {
+  text-transform: uppercase;
+}
+
+h2 {
+  font-size: 1.25rem;
+}
+@media (min-width: 431px) {
+  h2 {
+    font-size: 1.5rem;
+  }
+  #root {
+    padding: 2rem;
+  }
 }
