+++ src/index.css
@@ -1,11 +1,18 @@
 :root {
+  --color-body: rgb(255 255 255);
+  --color-block: rgb(245 245 245);
+  --color-black: rgb(17 17 17);
+  --color-accent: rgb(79 165 201);
+  --brightness-hover-: 0.8;
+  --break-point-sm: 769px;
+
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
   line-height: 1.5;
   font-weight: 400;
 
   color-scheme: light dark;
-  color: rgba(255, 255, 255, 0.87);
-  background-color: #242424;
+  color: var(--color-black);
+  background-color: var(--color-body);
 
   font-synthesis: none;
   text-rendering: optimizeLegibility;
@@ -25,7 +32,7 @@ a:hover {
 body {
   margin: 0;
   display: flex;
-  place-items: center;
+  justify-content: center;
   min-width: 320px;
   min-height: 100vh;
 }
